fix(js):detailed modules
