+++ package.json
@@ -22,6 +22,7 @@
     "express": "^4.16.4",
     "express-session": "^1.15.6",
     "http": "0.0.0",
+    "https": "^1.0.0",
     "path": "^0.12.7",
     "serve-static": "^1.13.2"
   }
